move Signable trait to solana-keypair
