+++ sdk/keypair/src/lib.rs
@@ -18,6 +18,7 @@ use {
 
 #[cfg(feature = "seed-derivable")]
 pub mod seed_derivable;
+pub mod signable;
 
 /// A vanilla Ed25519 key pair
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
