+++ sdk/keypair/src/signable.rs
@@ -1,0 +1,23 @@
+use {
+    crate::Keypair,
+    solana_pubkey::Pubkey,
+    solana_signature::Signature,
+    solana_signer::Signer,
+    std::borrow::{Borrow, Cow},
+};
+
+pub trait Signable {
+    fn sign(&mut self, keypair: &Keypair) {
+        let signature = keypair.sign_message(self.signable_data().borrow());
+        self.set_signature(signature);
+    }
+    fn verify(&self) -> bool {
+        self.get_signature()
+            .verify(self.pubkey().as_ref(), self.signable_data().borrow())
+    }
+
+    fn pubkey(&self) -> Pubkey;
+    fn signable_data(&self) -> Cow<[u8]>;
+    fn get_signature(&self) -> Signature;
+    fn set_signature(&mut self, signature: Signature);
+}
